+++ app/controllers/application_controller.rb
@@ -45,42 +45,6 @@ class ApplicationController < ActionController::Base
 
   private
 
-  def download_or_show_image(type, render_to, width)
-    html = image_html(type)
-
-    render html: html.html_safe and return unless render_to == 'jpg'
-
-    kit = build_kit(html, "#{type.tr('_', '-')}.css", width)
-
-    send_data(kit.to_jpg, type: 'image/jpg', filename: "#{type}.jpeg")
-  end
-
-  def image_html(image_type)
-    render_to_string(partial: image_type,
-                     locals: { app_config: @app_configuration, user: @user,
-                               render_to: params[:render_to]&.to_sym,
-                               context: params[:context]&.to_sym,
-                               company: Company.find_by_id(params[:company_id]) })
-  end
-
-  def build_kit(html, image_css, width)
-    kit = IMGKit.new(html, encoding: 'UTF-8', width: width, quality: 100)
-    kit.stylesheets << Rails.root.join('app', 'assets', 'stylesheets',
-                                       image_css)
-    kit
-  end
-
-  def set_app_config
-    # Need app config items for proof-of-membership
-    @app_configuration = AdminOnly::AppConfiguration.last
-  end
-
-  def allow_iframe_request
-    response.headers.delete('X-Frame-Options')
-    # https://stackoverflow.com/questions/17542511/
-    # cannot-display-my-rails-4-app-in-iframe-even-if-x-frame-options-is-allowall
-  end
-
   def set_locale
     @locale = I18n.default_locale
     @locale = params[:locale].to_s if params[:locale].present?
